add another p2 implementation
calling it the wobbling window. we're starting with a window of size 2
at position 0. we then shrink the window until the sum of the contents
is less the the target OR the size is 2. then we grow the window until
the sum of the contents are greater than or equal to the target OR there
are no more inputs remaining. then we quickly check if the contents of
the current window sum to our target, if so, find min and max in the
window and return. otherwise, advance the window forward and remove the
old first element from the sum and size.